# improve progress bar performance