--- a/toolbox/core/bst_progress.m
+++ b/toolbox/core/bst_progress.m
@@ -185,10 +185,10 @@ switch (lower(commandName))
             pBar.jProgressBar.setIndeterminate(1);
             pBar.jProgressBar.setStringPainted(0);
             % Set progress bar bounds
-            pBar.jProgressBar.setMinimum(0); pBar.Values.Minimum = 0;
-            pBar.jProgressBar.setMaximum(100); pBar.Values.Maximum = 100;
+            pBar.jProgressBar.setMinimum(0); GlobalData.Program.ProgressBar.Values.Minimum = 0;
+            pBar.jProgressBar.setMaximum(100); GlobalData.Program.ProgressBar.Values.Maximum = 100;
             % Set initial value to start
-            pBar.jProgressBar.setValue(0);  pBar.Values.Value = 0;
+            pBar.jProgressBar.setValue(0);  GlobalData.Program.ProgressBar.Values.Value = 0;
             
         % Call: bst_progress(''start'', title, msg, start, stop)
         elseif ((nargin == 5) && ischar(varargin{2}) && ischar(varargin{3}) && isnumeric(varargin{4}) && isnumeric(varargin{5}))
@@ -214,6 +214,9 @@ switch (lower(commandName))
             pBar.jProgressBar.setMaximum(valStop);  GlobalData.Program.ProgressBar.Values.Maximum = valStop;
             % Set initial value to start
             pBar.jProgressBar.setValue(valStart); GlobalData.Program.ProgressBar.Values.Value = valStart;
+
+            % Set last updated value: 
+            GlobalData.Program.ProgressBar.Values.LastUpdate = valStart;
             
         else
             error(['Usage : bst_progress(''start'', title, comment) ' 10 '        bst_progress(''start'', title, comment, valStart, valStop)']);
@@ -260,8 +263,13 @@ switch (lower(commandName))
         % Get the incremented progress bar position
         newVal = min(curValue + valInc, GlobalData.Program.ProgressBar.Values.Maximum);
         GlobalData.Program.ProgressBar.Values.Value = newVal;
-        pBar.jProgressBar.setValue(newVal);
         
+        
+        if newVal > ( GlobalData.Program.ProgressBar.Values.LastUpdate + ( GlobalData.Program.ProgressBar.Values.Maximum - GlobalData.Program.ProgressBar.Values.Minimum )/ 100)
+            pBar.jProgressBar.setValue(newVal);
+            GlobalData.Program.ProgressBar.Values.LastUpdate = newVal;
+        end
+
     % ==== SET POSITION ====
     case 'set'
         % Parse arguments