--- a/toolbox/core/bst_progress.m
+++ b/toolbox/core/bst_progress.m
@@ -153,7 +153,6 @@ if isempty(pBar)
     pBar.jWindow.setLocation(pos(1), pos(2));
 
     pBar.Values = struct('Minimum', 0, 'Maximum', 100, 'Value', 0, 'LastUpdate', 0);
-);
     % Save progress bar
     GlobalData.Program.ProgressBar = pBar;
 end