--- a/toolbox/core/bst_progress.m
+++ b/toolbox/core/bst_progress.m
@@ -151,8 +151,7 @@ if isempty(pBar)
     pos = [jLoc.getX() + ((jSize.getWidth() - DefaultSize.getWidth()) / 2), ...
            jLoc.getY() + ((jSize.getHeight() - DefaultSize.getHeight()) / 2)];
     pBar.jWindow.setLocation(pos(1), pos(2));
-
-    pBar.Values = struct('Minimum', 0, 'Maximum', 100, 'Value', 0, 'LastUpdate', 0);
+    pBar.Values = struct('Minimum', [], 'Maximum', [], 'Value', [], 'LastVal');
     % Save progress bar
     GlobalData.Program.ProgressBar = pBar;
 end
@@ -185,10 +184,15 @@ switch (lower(commandName))
             pBar.jProgressBar.setIndeterminate(1);
             pBar.jProgressBar.setStringPainted(0);
             % Set progress bar bounds
-            pBar.jProgressBar.setMinimum(0); GlobalData.Program.ProgressBar.Values.Minimum = 0;
-            pBar.jProgressBar.setMaximum(100); GlobalData.Program.ProgressBar.Values.Maximum = 100;
+            pBar.jProgressBar.setMinimum(0);
+            pBar.jProgressBar.setMaximum(100);
             % Set initial value to start
-            pBar.jProgressBar.setValue(0);  GlobalData.Program.ProgressBar.Values.Value = 0;
+            pBar.jProgressBar.setValue(0);
+            % Update values in GlobalData
+            GlobalData.Program.ProgressBar.Values.Minimum = 0;
+            GlobalData.Program.ProgressBar.Values.Maximum = 100;
+            GlobalData.Program.ProgressBar.Values.Value   = 0;
+            GlobalData.Program.ProgressBar.Values.LastVal = 0;
             
         % Call: bst_progress(''start'', title, msg, start, stop)
         elseif ((nargin == 5) && ischar(varargin{2}) && ischar(varargin{3}) && isnumeric(varargin{4}) && isnumeric(varargin{5}))
@@ -210,14 +214,14 @@ switch (lower(commandName))
                 valStop  = 100;
             end
             % Set progress bar bounds
-            pBar.jProgressBar.setMinimum(valStart); GlobalData.Program.ProgressBar.Values.Minimum = valStart;
-            pBar.jProgressBar.setMaximum(valStop);  GlobalData.Program.ProgressBar.Values.Maximum = valStop;
-            % Set initial value to start
-            pBar.jProgressBar.setValue(valStart); GlobalData.Program.ProgressBar.Values.Value = valStart;
-
-            % Set last updated value: 
-            GlobalData.Program.ProgressBar.Values.LastUpdate = valStart;
-            
+            pBar.jProgressBar.setMinimum(valStart);
+            pBar.jProgressBar.setMaximum(valStop);
+            pBar.jProgressBar.setValue(valStart);
+            % Update values in GlobalData
+            GlobalData.Program.ProgressBar.Values.Minimum = valStart;
+            GlobalData.Program.ProgressBar.Values.Maximum = valStop;
+            GlobalData.Program.ProgressBar.Values.Value   = valStart;
+            GlobalData.Program.ProgressBar.Values.LastVal = valStart;
         else
             error(['Usage : bst_progress(''start'', title, comment) ' 10 '        bst_progress(''start'', title, comment, valStart, valStop)']);
         end
@@ -261,16 +265,18 @@ switch (lower(commandName))
             error('Usage : bst_progress(''inc'', valInc)');
         end
         % Get current value
+        minValue = GlobalData.Program.ProgressBar.Values.Minimum;
+        maxValue = GlobalData.Program.ProgressBar.Values.Maximum;
         curValue = GlobalData.Program.ProgressBar.Values.Value;
-        % Get the incremented progress bar position
-        newVal = min(curValue + valInc, GlobalData.Program.ProgressBar.Values.Maximum);
-        GlobalData.Program.ProgressBar.Values.Value = newVal;
-        
-        
-        if newVal > ( GlobalData.Program.ProgressBar.Values.LastUpdate + ( GlobalData.Program.ProgressBar.Values.Maximum - GlobalData.Program.ProgressBar.Values.Minimum )/ 100)
+        newVal = min(curValue + valInc + minValue, maxValue);
+        % Plot the incremented progress if it moves at least 1% of the bar range
+        if (abs(newVal - GlobalData.Program.ProgressBar.Values.LastVal) / (maxValue - minValue)) > 1/100
+            % Get the incremented progress bar position
             pBar.jProgressBar.setValue(newVal);
-            GlobalData.Program.ProgressBar.Values.LastUpdate = newVal;
+            % Update value in GlobalData
+            GlobalData.Program.ProgressBar.Values.LastVal = newVal;
         end
+        GlobalData.Program.ProgressBar.Values.Value = newVal;
 
     % ==== SET POSITION ====
     case 'set'
@@ -282,12 +288,13 @@ switch (lower(commandName))
         end
         % Get current value
         curValue = GlobalData.Program.ProgressBar.Values.Value;
-        % Get the incremented progress bar position
+        % Plot the position if it changes
         if (curValue ~= newVal)
             newVal = min(newVal, GlobalData.Program.ProgressBar.Values.Maximum);
-            GlobalData.Program.ProgressBar.Values.Value = newVal;
-            GlobalData.Program.ProgressBar.Values.LastUpdate = newVal;
             pBar.jProgressBar.setValue(newVal);
+            % Update value in GlobalData
+            GlobalData.Program.ProgressBar.Values.Value   = newVal;
+            GlobalData.Program.ProgressBar.Values.LastVal = newVal;
         end
         
     % ==== GET POSITION ====