@rcassani 2025-09-19
@Edouard2laire, can you test this on your Mac?
The performance is also improved on Linux though it is not as drastic, (`main1` from 2s to 1s)