@Edouard2laire 2025-09-19
tested, it seems to be working. the time is now around .5 seconds.